fix: throw exception for listener and delete payloads

+++ event-notificator-service/src/main/java/git/xlamid/eventnotificatorservice/notification/service/EventNotificationService.java
@@ -5,10 +5,13 @@ import git.xlamid.eventnotificatorservice.exception.model.validation.EmptyValida
 import git.xlamid.eventnotificatorservice.notification.dto.GetNotificationDto;
 import git.xlamid.eventnotificatorservice.notification.dto.MarkNotificationDto;
 import git.xlamid.eventnotificatorservice.notification.entity.NotificationEntity;
+import git.xlamid.eventnotificatorservice.notification.entity.NotificationEventPayloadEntity;
 import git.xlamid.eventnotificatorservice.notification.mapper.NotificationMapper;
+import git.xlamid.eventnotificatorservice.notification.repository.NotificationEventPayloadRepository;
 import git.xlamid.eventnotificatorservice.notification.repository.NotificationRepository;
 import git.xlamid.eventnotificatorservice.security.service.UserSecurityContextService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
@@ -17,11 +20,13 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.OffsetDateTime;
 import java.util.List;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class EventNotificationService {
 
     private final NotificationRepository notificationRepository;
+    private final NotificationEventPayloadRepository payloadRepository;
     private final NotificationMapper notificationMapper;
 
     private final UserSecurityContextService userSecurityContextService;
@@ -63,6 +68,18 @@ public class EventNotificationService {
         OffsetDateTime threshold = OffsetDateTime.now().minusDays(7);
         List<NotificationEntity> notificationEntities = notificationRepository
                 .findAllOldAndReadNotifications(threshold);
-        notificationRepository.deleteAll(notificationEntities);
+        List<NotificationEventPayloadEntity> payloadEntities = payloadRepository
+                .findAllByNotificationsIsNull();
+
+        if (notificationEntities != null && !notificationEntities.isEmpty()) {
+            List<Long> notificationIds = notificationEntities.stream().map(NotificationEntity::getId).toList();
+            log.info("Clearing old notifications: {}", notificationIds);
+            notificationRepository.deleteAll(notificationEntities);
+        }
+        if (payloadEntities != null && !payloadEntities.isEmpty()) {
+            List<Long> payloadIds = payloadEntities.stream().map(NotificationEventPayloadEntity::getId).toList();
+            log.info("Clearing old payloads: {}", payloadIds);
+            payloadRepository.deleteAll(payloadEntities);
+        }
     }
 }
